Added GPT4 Mobile V2 Support
